requirements.txt
python-docx==0.8.11
mammoth==1.6.0
textract==1.6.5 